refactor(xxs): move errors to pkg errors
