+++ dirtools/dirtools.go
@@ -2,6 +2,7 @@ package dirtools
 
 import (
 	"fmt"
+	"github.com/therealkevinard/gitdir/errors"
 	"io/fs"
 	"net/url"
 	"os"
@@ -9,8 +10,6 @@ import (
 	"path/filepath"
 	"regexp"
 	"strings"
-
-	"github.com/therealkevinard/gitdir/commandtools"
 )
 
 // regexes.
@@ -50,7 +49,7 @@ func NormalizeRepoURL(repoURL string) (string, error) {
 		return "", fmt.Errorf("error parsing url: %w", err)
 	}
 	if parsed.Host == "" || parsed.Path == "" {
-		return "", fmt.Errorf("invalid url %s (sanitized to %s): %w", repoURL, cleanRepoURL, commandtools.ErrInvalidURL)
+		return "", fmt.Errorf("invalid url %s (sanitized to %s): %w", repoURL, cleanRepoURL, errors.ErrInvalidURL)
 	}
 
 	// build final directory
